+++ grass7/raster/r.keras/r.keras.predict/r.keras.predict.py
@@ -101,12 +101,12 @@ def main():
     print(class_raster.shape)
     y_train = class_raster[class_raster > 0]
     print(y_train.shape)
-    x_train_input = np.zeros((y_train.shape[0], len(input_files)), dtype=np.float32)
+    x_train_input = np.zeros((y_train.shape[0], len(input_files)), dtype=np.float16)
     print(x_train_input.shape)
     for i in range(len(input_files)):
         x_train_input[:,i] = getData(input_files[i])[class_raster > 0]
 
-    x_train_subinput = np.zeros((y_train.shape[0], len(subinput_files)), dtype=np.float32)
+    x_train_subinput = np.zeros((y_train.shape[0], len(subinput_files)), dtype=np.float16)
     for i in range(len(subinput_files)):
         x_train_subinput[:,i] = getData(subinput_files[i])[class_raster > 0]
